+++ NeuroSharp/Layers/ParameterizedLayer.cs
@@ -5,9 +5,9 @@ namespace NeuroSharp
 {
     public abstract class ParameterizedLayer : Layer
     {
-        public Matrix<float> Weights { get; set; }
-        public Vector<float> Bias { get; set; }
-        public Matrix<float> WeightGradient { get; set; }
-        public Vector<float> BiasGradient { get; set; }
+        public Matrix<double> Weights { get; set; }
+        public Vector<double> Bias { get; set; }
+        public Matrix<double> WeightGradient { get; set; }
+        public Vector<double> BiasGradient { get; set; }
     }
 }
